+++ lib/models/launchConfig.dart
@@ -4,6 +4,7 @@ class LaunchConfig {
   String? server;
   String? static;
   String? playerId;
+  String? playerToken;
   String? launchType;
   String? gameContainerId;
   bool? rawUrl;